catch and descriminate different exception types
This change catches the different exceptions types thrown by the C++ SDK
and throws a JS exception with the property `type`, which is the
discriminator, and the property `exception`, which is an actually
JavascriptError as one would expect.

For the record, I am not proud of the macros used here :).